+++ threedigrid_builder/grid/embedded.py
@@ -172,6 +172,10 @@ class EmbeddedObjects:
     @classmethod
     def from_cell_tree(cls, objects, cell_tree):
         """Construct the velocity points in the embedded objects from a cell tree."""
+        # cell_tree.query_bulk(
+        #     pygeos.get_point(objects.the_geom, 0),
+        
+
         # The query_bulk returns 2 1D arrays: one with indices into the objects
         # and one with indices into the cells.
         idx = cell_tree.query_bulk(objects.the_geom, "intersects")
@@ -248,22 +252,17 @@ class EmbeddedObjects:
         )
         vp_idx_a, cell_idx_a = cell_tree.query_bulk(pnt_a)
         vp_idx_b, cell_idx_b = cell_tree.query_bulk(pnt_b)
-        # List velocity points that are at the model edge. These are the vpoints that
-        # occur only in vp_idx_a (before) or only in vp_idx_b (after)
-        at_model_edge = np.setxor1d(vp_idx_a, vp_idx_b)
-        if len(at_model_edge) > 0:
-            # remove the velocity points at model edge:
-            self.vpoint_s = np.delete(self.vpoint_s, at_model_edge)
-            self.vpoint_ch_idx = np.delete(self.vpoint_ch_idx, at_model_edge)
 
-            # adapt the return of the bulk query correspondingly:
-            mask_a = ~np.isin(vp_idx_a, at_model_edge)
-            cell_idx_a = cell_idx_a[mask_a]
-            mask_b = ~np.isin(vp_idx_b, at_model_edge)
-            cell_idx_b = cell_idx_b[mask_b]
-            idx_mapping = np.insert(np.arange(len(self.vpoint_s)), at_model_edge, -9999)
-            vp_idx_a = idx_mapping[vp_idx_a[mask_a]]
-            vp_idx_b = idx_mapping[vp_idx_b[mask_b]]
+        # List velocity points that are outside of the model or at the model edge.
+        n_vpoints = len(self.vpoint_s)
+        no_intersct_a = np.bincount(vp_idx_a, minlength=n_vpoints) == 0
+        no_intersct_b = np.bincount(vp_idx_b, minlength=n_vpoints) == 0
+        if np.any(no_intersct_a | no_intersct_b):
+            bad_objects_ids = self.vpoint_ch_idx[no_intersct_a | no_intersct_b].tolist()
+            raise SchematisationError(
+                f"{self.objects.__class__.__name__} {sorted(set(bad_objects_ids))} are "
+                f"not completely inside the 2D cell."
+            )
 
         cell_idx_a, cell_idx_b = self._fix_tangent_lines(
             vp_idx_a, cell_idx_a, vp_idx_b, cell_idx_b